fixing issue with delete image/product

+++ util/file.js
@@ -1,11 +1,13 @@
 const fs = require('fs');
 
 const deleteFile = (path) => {
-    fs.unlink(path, err => {
-        if (err) {
-            throw(err);
-        }
-    });
+    if (fs.existsSync(path)) {
+        fs.unlink(path, err => {
+            if (err) {
+                throw(err);
+            }
+        });
+    }
 }
 
 exports.deleteFile = deleteFile;
